fix(git): grant the work tree root rather than the common git dir's parent
