+++ README.md
@@ -13,7 +13,7 @@ The sandbox uses [bubblewrap](https://github.com/containers/bubblewrap) on Linux
 - **Allowed packages**: the binaries you list in `allowedPackages` are on the agent's PATH, together with `bash` and `cacert`.
 - **Network**: open internet access by default, with host-local services blocked. Set `allowedDomains` to limit the internet domains. Set `allowedLocalPorts` to permit specific host-local TCP ports.
 - **Environment**: only the variables you pass through `env` reach the agent. The sandbox clears the rest of the host environment.
-- **Git**: read/write access to the git directory, and read-only access to the directory that contains it. The same applies when the git directory is outside the project tree (worktrees). See [Git](#git).
+- **Git**: read/write access to the git directory, and read-only access to the working tree you launched in. See [Git](#git).
 - **Nix**: disabled by default. You can let the agent run nix commands.
 
 The sandbox denies everything else. `$HOME` is an ephemeral writable tmpfs that disappears when the sandbox exits. There is one exception. If you launch the agent from your home directory, the sandbox exposes that directory read-write, like any other launch directory. See [Launching from your home directory](#launching-from-your-home-directory).
@@ -300,20 +300,26 @@ Local git operations work with no extra configuration. The agent can switch bran
 
 ### What the sandbox exposes
 
-At launch, the sandbox asks git for the common git directory. Git searches upward from the launch directory to find it. If there is no repo, the sandbox exposes no git paths. If there is a repo, the sandbox exposes three paths:
+At launch, the sandbox finds two directories. It asks git for the common git directory. It then searches upward from the launch directory for the nearest `.git` entry, which gives the working tree root. If there is no repo, the sandbox exposes no git paths.
+
+If there is a repo, the sandbox exposes these paths:
 
 | Path | Access |
 |---|---|
 | The launch directory | read-write |
-| The git directory | read-write, except the [read-only paths](#read-only-paths-in-the-git-directory) |
-| The parent of the git directory | read-only |
+| The .git directory | read-write, except the [read-only paths](#read-only-paths-in-the-git-directory) |
+| The working tree root | read-only, and only when it is not the launch directory |
 
-For an ordinary repo, that parent is the repo root. This is why git works when you launch from a subdirectory. Two cases are different:
+The working tree root is what lets git report on files above the launch directory. Without it, `git status` and `git diff` report those files as deleted.
 
-- **Worktrees:** the git directory belongs to the main repo, so its parent is the main repo root. The agent can read the main checkout, and each sibling worktree below it.
-- **Submodules:** the git directory is `<superproject>/.git/modules/<name>`, so its parent is `.git/modules`. The sandbox does not expose the superproject working tree. The agent can read the git directory of other submodules.
+The sandbox exposes no other directory. It permits traversal of the directories that lead to these paths, so git can reach the git directory. The agent cannot read the files in a directory it can only traverse.
 
-If that parent is your home directory or above it, the sandbox disables git for the session and prints a warning. A repo whose root is `$HOME` therefore gets no git support. The exception is a launch from `$HOME` itself.
+The working tree root is the root of the tree you launched in. It is not the root of the repo above it:
+
+- **Worktrees:** the working tree root is the worktree itself. The sandbox does not expose the main checkout, or any sibling worktree.
+- **Submodules:** the working tree root is the submodule itself. The sandbox does not expose the superproject working tree, or the git directory of any other submodule.
+
+The sandbox disables git for the session when the directory that contains the git directory is your home directory or above it. It prints a warning when this happens. A repo whose root is `$HOME` therefore gets no git support. The exception is a launch from `$HOME` itself.
 
 ### Remote access (push / pull / fetch)
 
@@ -519,7 +525,7 @@ The sandbox refuses a launch directory above `$HOME` (`/`, `/home`, `/Users`) ou
 - A launch from `$HOME` turns off home masking entirely. See [Launching from your home directory](#launching-from-your-home-directory). Everything below assumes that you launch from a project directory.
 - Your username and home directory path are visible to the agent. This is unavoidable, because the agent needs to know where `$HOME/.claude` resolves to. If your username is itself sensitive, this is not the right tool.
 - All of `/nix/store` is readable, not only your allowed packages. The allowlist restricts execution only. The Nix store is normally world-readable on any system, so this matches existing behavior, but it does mean that the agent can list every package you have built.
-- A launch from a subdirectory does not limit reads to that subdirectory. The agent can read the whole repo. From a worktree, it can also read the main checkout. See [What the sandbox exposes](#what-the-sandbox-exposes).
+- A launch from a subdirectory does not limit reads to that subdirectory. The agent can read the whole working tree that contains it. See [What the sandbox exposes](#what-the-sandbox-exposes).
 - The agent can read all of the git directory. This includes every branch, stash and reflog entry, also content that is no longer in the working tree.
 
 ### Linux vs macOS
