+++ tests/fixtures/git-topologies.nix
@@ -1,10 +1,12 @@
-# Test fixture: sandbox with exposeRepoRoot = true
+# Test fixture: a sandbox with git, for the four launch topologies (repository
+# root, repository subdirectory, worktree/submodule root, worktree/submodule
+# subdirectory).
 { pkgs ? import ../pinned-nixpkgs.nix { } }:
 let
   sandbox = import ../../default.nix { pkgs = pkgs; };
 in sandbox.mkSandbox {
   pkg = pkgs.bashInteractive;
   binName = "bash";
-  outName = "sandboxed-bash";
+  outName = "sandboxed-bash-git-topologies";
   allowedPackages = [ pkgs.coreutils pkgs.git ];
 }
